ldap: Implement LDAP support.
This happens transparantly for the rest of the application. We push all
information unidirectionally into the LDAP and do not read from it. This is
suprisingly easy to implement once the LDAP basics have settled in the
brain.

With regard to passwords: they are stored in salted and hashed
form. Verification is done with the tool ldapwhoami and a trial bind thus
checking the password.

Conflicts:
	cdedb/backend/core.py