cdedb: Add auto-build.
This adds an auto-build, so we can use a standardized vm-image for
development. The idea is for this to be as close to the production
environment as possible. Some adjustments are necessary for compatibility
with Debian and the settings in the target environment.

The initial auto-build script was created by Helmut Grohne and Klaus Aehlig
and later adapted to CdEDB usage by Dominik Brodowski.