cdedb: refactor changelog functionality
The core backend has to have access to the changelog, otherwise we will have
major troubles in the future when we want to change the stati of members and
such.

Unfortunately all solutions have some drawbacks. This one muddles the waters
by moving functionality into the core realm which does not entirely belong
there. But this seem to be the least unpleasant one.