cdedb: Add account creation.
This adds creation of accounts by admins, as well as anonymous requests for
accounts (which are moderated by admins). It also includes a major cleanup
of whitespace before colons.

Also now the event_user and assembly_user roles are above the ml_user role
to allow them access to the mailing lists.