event-frontend: Implement event management.
This adds most functionality needed for event management (like registration,
...). One big chunk still missing is the offline support, but this is
orthogonal to the current change.

We now use JQuery UI for some interactivity. Additionally we now harness
LaTeX for document generation. For the latter we need an update to the VM
image.

Furthermore we now generate performance information (in form of time take to
process a request).

Last, we have a new URL scheme, that is more RESTful in that it has a more
hierarchical structure. This has the upshot of being much more predictable
than the old scheme (in terms of what should the URL be if I want to do
foo?).