assembly-backend: Introduce the assembly backend.
This does pretty much what the headline says. We implement a preference
voting system with the Schulze method as evaluation. The system is designed
such that the voting secrets can be externalised and have never to be stored
on the server. However for usability we store them until the corresponding
assembly is concluded.

This also makes use of the previous commit allowing realms without
additional user attributes. Now it should be rather easy and fast to add a
new such realm.