frontend: Add ambience context
This automatically resolves (nearly) all object ids in the URL path into the
corresponding data. This would be done by the handler anyway (in most
cases), so we don't incur any additional cost, but reduce explicit calls.

Additionally this defines standard way to pass data to the templates, so
that the navigation can use it. We can detect e.g. the presence of an
ambient event and incorporate its details into the navigation.

However this revealed some rough edges: this is broken for past_events until
they migrate from event_id to past_event_id, assembly attachements are
solved somewhat non-canonical and this has some side-effects on the
test-suite.