cdedb: remove RPC mechanism.
This replaces the RPC calls by direct invocations. This has several
advantages.

(1) Allow usage of the Atomizer in the Frontend. Thus we can make consistent
    changes, which was previously not possible.

(2) Reduced complexity which makes everything easier. (This kills about 800
    lines of code.)

(3) As an unimportant side-effect, everything should be faster now.

The downside is a less paranoid setup and the possibility to circumvent the
frontend/backend separation (this is forbidden by convention, so shouldn't
be an issue).