changelog: Improve changelog semantics.
We previously compared against the pending revision to find out if changed
fields require a review. This is nonsense since we need to compare against
the committed version. We additionally increased the robustness by making
any change superseeding a pending change also require review.

At the same time we use the pending data in the user change forms.