UI/Core: Highlight uncommited versions in profile history view.
Unfortunately, there's a minor glitch in this highlighting:
If a pending version gets superseded by another change, fields that only
changed in the first changeset are shown as 'superseded', but should be
shown in the color of the next committed or nacked version. Fixing this
would need some complicated python preprocessing; inheriting the state
of a field's version from the next non-superseded version, but only if
no change happened between these versions.