core-backend: Change return code of user editing if no change was made.
This could introduce spurious errors if we want to change to the already
existing state. Semantics here are debateable: Since we didn't change
anything it may be a failure, but on the other hand the endstate is the
desired state, so that it may be a success. We go the easier route and count
it as a success.