event: Add customizable fields to lodgements and courses.
Often the need arises for some additional data to be attached to lodgements
or courses. We provide a notes field for this purpose, but many times it is
desireable to add this data in a more structured way. These fields now allow
exactly this.

This does the work in the backend and partial frontend. The styling and JS
interaction are not yet done.

This obviously changes the DB schema, causing the need for an update.