Revert "Fix event/register and event/amend_registration: Allow chosing of courses without active segment"
This reverts commit 901ce6edd0918431782ae75d15abe80611ce8874.

As discussed in recent e-mails, the 'active_segments' feature can be
used by orgas for two different purposes:

* Flag courses that have been cancelled by the instructor while
registration is open, to prevent further participants from applying for
this course.
* Flag courses that have been cancelled by orgas due to limited
ressources while assigning course attendees.

For the first purpose it is helpful if the course is no longer
selectable by participants. For the second purpose it doesn't matter,
since assigning of attendees is typically done *after* the registration
period. (Here, it's merely important, that cancelled courses are still
selectable in the orga interface to avoid losing information about
course choices.