event: Prevent part/track creation and deletion if registrations are present.
Otherwise we would have to do some major lifting to keep these operations
functional. And they should only be required at the beginning when an event
is set up.

The plan is to add functionality to delete registriations, which takes away
the one remaining bump with this (orgas creating a test registration and
then being unable to create/delete a part/track).