Add preliminary version of mailinglist software.
This is a preview version which should already be very close to the
interface, the final thing will have. So this should enable us to develop
all the handling of our side while in parallel the mailinglist software
matures.

Big thangs to Roland Köbler for developing the mailinglist software. :)