frontend: Pass query form choices as sorted list to JavaScript
This closes #150

I zapped AbstractFrontend.enum_choice() as this would be (nearly) a
duplicate of enum_entries_filter(). JavaScript/JSON objects are not
sorted (old JS standard) or sorted by strange criteria (since ECMA
Script 2015), so we need lists for the sorted select boxes. For filling
result lists, we still need a dict-like object to look up the values.
Thus, we construct choices as OrderedDicts and pass them twice to the
template.