JS: Fix query queryform.js: Allow search after download
With 7dfb524 we introduced a hidden input which is added dynamically on
submit button click to pass the submit button's value to the server,
although our custom submit handler (which shortens the query URL). But
it hasn't been removed after the submit, so the normal "Search" button
didn't work, if used after one of the download buttons.