backend: Rework relative admin permission model
This commit mainly rewrites the common.privilege_tier() function. It is
now based upon an explicit defined realm hierarchy
(`common.REALM_INHERITANCE`) and follows a slightly modified logic: By
now, an admin was able to edit another user, only if they were a admin
of the (single) highest realm in the hierarchy the user was member of.
With this change, a user can be edited by admins of any its realms, which
is not implied by another of its realms.

Examples:
* User in {ml, event} can be edited by event_admins (not ml_admins,
  since event implies ml)
* User in {ml, event, assembly} can be edited by event_admins as well
  as assembly_admins (since both realms are independend)
* User in {ml, event, assembly, cde} can only be edited by cde_admins
  (since cde implies all other realms)
(of course, all of them can be edited by core_admins)

This finally fixes #110.

Additionally we can use the explicit declaration of the realm hierarchy
in the core/promote_user routine.