core: Improve privilege implementation.
This does two things: first we unify privilege_tier and
check_create_privilege into one function with consistent interface; second
we fix the logic in the core frontend w.r.t. which admin can edit which
attributes (previously an assembly admin could edit the event fields of a
persona in the appropriate constellation).