frontend/core: Show email in select_persona if searched for it
This modification makes sure to transmit the email address of result
data sets of the select_persona API if it matches one of the search
terms. Otherwise the result entry would fail the client-side filtering
of selectize.js.

To preserve maximum privacy, we only allow that special treatment, if
the relevant search term has a minimum length and contains the '@'
character.