Make sample data valid
This commit changes the postal codes in the sample data so that user data
can be changed without looking up valid postal codes before.

Lisa is still missing a birth date and can not be modified without
specifying one, but since she is supposed to have whacked data this change
does not do anything about that.

Closes #170.