Add calculated is_course_instructor column to registration query
This column is dynamically calculated by comparing the course_id to the
course_instructor columm. It will return NULL, if one of those is NULL,
TRUE, if they are equal (the participant is in the course, they are
course instructor of) or FALSE if they offered to bee course instructor,
but are assigned to another course.

This fixes #153.