core: Revamp show_user w.r.t. displayed data.
The new model is that each realm has its associated attributes with the
exception of some basic ones which are reserved for the core realm (like
is_archived).

Now for each realm R one needs some reason to be able to see the
corresponding attributes from the following list:

* the data is your own data
* you are core admin
* you are a relative admin w.r.t the dataset and admin in R
* for R in {ml, event}: you are moderator/orga and the dataset is linked to
  your list/event
* for R == cde: you use the member search

There are some minor special attributes with no big significance.