ui: Change button label/placeholder on event/show_event to fix layout
Fixes #237

With the previous german labels, the "add orga" button breaks the layout
at certain screen widths (e.g. at 1024px). This is caused by the way,
selectize.js dynamically sets the input field's width based on the
placeholder, which is not easily fixable. So the best fix was to shorten
button label and placeholder.