core: Replace password strength logic with basic zxcvbn
This commit replaces the basic password strength logic by using
zxcvbn, which tries to determine the actual entropy of a password
using different dictionaries, searching for patterns etc.

This change adds the possibility to provide user information like
name, email and birth date to _password_strength to spot usages of
these in passwords as well. Furthermore, this change makes sure that
errors raised by the _password_strength function are actually
displayed to the user.