frontend: Fix error handling and add error messages
HTTPExceptions must be risen to be displayed as a nice error page (using
Application.make_error_page). If they are returned, they are interpreted
as the WSGI reposonse and displayed as a minimalistic white page with
the error description.

Additionally, now we have internationalized error descriptions. (Note,
that the error descriptions are internationalized upon generation, not
in the error template, to allow insertion of dynamic information.)