core: Add mock German dictionary for zxcvbn
In theory, with this change zxcvbn should conside GERMAN_WIKIPEDIA_WORDS
as additional ranked dictionary. In practice, I could confirm that it
is possible to override given dictionaries with this but was not able
to make zxcvbn reject words of this list. It might be possible that
this resolves itself once a more useful list is provided, but I doubt
it.