security: Add word counts from German Wikipedia
This adds a statistic about word counts in German Wikipedia as well
as a Python file containing the first 30000 entries of these list in
an ordered fashion.

To generate this, you download a dewiki-latest-pages-articles.xml.bz2
dump from https://dumps.wikimedia.org/dewiki/latest/ and run
attardi/WikiExtractor (with the params specified in the second script).
This required about 26 GB of free disk space and 10 hours of processing
time, but runs in parallel. Afterwards, zxcvbn/data-scripts/count_wikipedia.py
is run on this directory, which required another 10 hours to generate
the word count. The second scipt could probably be heavily optimized.

Sadly, the second script (count_wikpedia.py) bundled with zxcvbn is
not that well made and has large problems with umlauts and other
non-unicode chars, sometimes they are omitted, while at other times,
they are replaced with multiple other chars, e.g. ö by ap.

Furthermore, in some cases words are seemingly cut off prematurely
for unknown reasons, probably related to encoding as well (e.g. "fa").
It would probably not be too hard to fix the script to work just
fine, but redoing this requires quite much time. The WikiExtractor
outpur still exists and could be reused. I can provide parts of it
to test a fixed script. (As a whole it is larger than 5 GB.)