core-backend: Harden doppelganger detection against resource exhaustion.
Previously this exhibited an O(n^2) behaviour if many similar persons
existed (as in the case of the load testing). We now limit the number of
matches to achieve an O(n) behaviour.

For practical matters this should even enhance the usability, since we now
sort for match fitness and present only the best matches. Dropping something
relevant seems very unlikely.

Closes #282