Add check_escaping.py script and accompanying sample data set
This script provides a way to do some fuzzy searching for missing HTML
escaping in the frontend/templates. It uses the marker
"<script>abcdef</script>" to find unescaped and double-escaped HTML
code, which is served from the database (persistent XSS), see sample
dataset, or from the HTTP request (reflected XSS).

Refs #255