frontend: Move Anti-CSRF check to application and change behaviour
Instead of wrapping all POST endpoint handlers, which require a CSRF
check, we mark them with the check_anti_csrf attribute, and let the
Application do the check before calling the handler function. We also
changed the behaviour on failure from rising a 403 exception to adding
validation errors (and a notification) to allow the user to resubmit
the form.

The RequestState's new csrf_alert property can be used later to disallow
any database state modifications in the backend, when a CSRF has been
detected.