frontend: Add Content-Security-Policy header
This header provides a whitelist for scripts, stylesheets and emebedable
objects, which (currently) only allows everything hosted on the
cdedb's domain, as well as inline-scripts and -styles.

This may be too hard (regarding embedding of images, which could by done
by reST/markdonw) on the one hand and to weak (regarding injection of
<script> tags) on the other hand. The latter one could be improved by
using 'nonce-...' instead of 'unsafe-inline', which unfortunately is
only supported by recent browser versions.

This is part of the XSS mitigation (see #255).