frontend: Add markdown Treeprocessor to reduce heading level
This mimics the behaviour we configured for the reST-parser: The highest
markdown heading level creates a `<h4>` tag, a second level heading
results in a `<h5>` tag.

Unfortunately the way of registring processors changed from
python-markdown version 2.6 to version 3.1. This requires a workaround
to make the code backwards compatible with Debian Stretch.