event-backend: First draft of partial-import functionality.
This is the minimal viable version without any amenities. It offers an
export and a symmetric import. The import touches only changed attributes.

It can create and delete things, however this should not be done with any
concurrent changes (as the ordering is an implementation detail that might
change). The caller is responsible for ensuring that deletions are possible
and do not violate referential integrity.