frontend/event: Add `skip` parameter to change_registration()
The `skip` URL parameter allows to specify fields that should not be
evaluated when submitting the change_registration form. They are also
rendered as readonly inputs (currently only implemented for
`reg.checkin`).

This fixes the Checkin workflow: By disabling the `reg.checkin` field,
(when accessing the form from the checkin page) we make sure that
submitting changes to the registration does not override the checkin
state, regardless of the order of actions.

fixes #416