scripts: Add mechanism for detecting changes needing adjustments in production.
Main example are probably SQL schema changes. If they are not manually
introduced in the production database the corresponding new code will simply
fail.

To mark such changes a special line is to be inserted in the commit
message. This line has to start with 'Deploy:' and further contain a short
description of the required change.

To demonstrate this we add the following line which is not caused by this
commit, but by e954778b. This commit currently sits in master but not in
stable, so this is actually a fix.

Deploy: The 'event_logo' directory must be created.