backend: Include table name in result column names of generic query
By default PostgreSQL only uses the name of the SQL column as result
column name, even if the SELECT clause included a complex column
specification like `table.column`. In combination with the
RealDictCursor of psycopg, this results in problems with multiple
columns with the same name and requires to use those ugly alias names in
the registration query.

However, it can be solved by giving an explicit alias name for the
result column. The alias may even include periods, so we can name the
result column exactly like the column name in our query spec.

This fixes #519 and my be first step towards #165. It also makes #513
less important, since the automatic replacement of course ids with
course names is fixed as a side effect.