bin: add script to upload old assembly files.
This creates a "Sonstiges" Assembly if one is not already present, enters (almost) all attachments into the database aswell as the storage.

Some afterthought will have to be given to zip attachments. (There are 3 of those.)

Deploy: The cdefiles folder needs to be at /cdedb2/bin/cdefiles unpacked,
so that /cdedb2/bin/cdefiles/downloads/[id].pdf is the path to the attachment with the given id.
Then run the script from any location. Has to have privileged access.