core/genesis: Move zip code test where it can always be done
If there is missing required data in the val dict, it is set to
None by the _examine_dictionary_fields function. That is why an
exception is thrown if, for example, adress is missing.

This commit fixes this by doing the Zip code check before the
required data check, which also allows for displaying this
error in any case the Zip code is problematic.