core: introduce a two-step, four-eye process for changing a personas admin bits.
This requires any change to a personas admin bits to be reviewed by a second superadmin before committing it. Therefore we require a second superadmin in the sample dataset. To avoid conflict with any existing users used in tests, we create a new superadmin for this purpose.

Deploy: Add the new table.