frontend/log: Fix count of log entries
The second number is now the count of log entries -1,
as the first log entry is counted as 0. I decided to
do subtract from the length instead of shifting the
first number to 1, so the numbers in the heading match
the internal entries id you can search for.

I do not think that this is the optimal way to handle
these numbers, for more information see #379. However,
this does at least fix the count of entries as well as
the search by entries, while not making things worse
for the other filters.

Closes #609