ml/frontend: Implement modification of own subscription state
This commit allows users to modify their own subscription state
according to our new data model. It fixes the show_mailinglist
page along the way.

A new may_subscribe function is introduced, which pretends to
give you the SusbcriptionPolicy of the mailinglist, but uses
audience end assembly attendance information to tinker with
this information to tell you what YOU can do with the mailinglist.
In a followup, event registration stati have to be respected
here as well.

Large parts of the functions implemented here should probably be
shifted to the backend later.