ml/test: Add thorough tests for some moderation policies
This adds thorough tests for opt-in, moderated opt-in and
mandatory mailinglists and fixes issues encountered along
the way.

As these tests require interactions of users, only admins
and moderators handling their own requests are explicitly
tested, as they can interact with a list as normal user and
as list manager. This should not be too bad, as these users
are not special cased anymore if they are using the same
interfaces as anybody else.

This commit drops any possibility to unsubscribe from
mandatory mailinglists, even by Unsubscribe Overrides,
as this is favorable for legal reasons and less problematic
from an architecture point of view. (Without special casing,
removing unsubsribe overrides here would leave these users
as explicitly unsubscribed, which is not tolerable in this
specific case.)