Add deployment hint forgotten earlier.
This should have been in commit
43321a52378bdb1ca048d5c9513be0aa3466a147.

Deploy: Add varchar column registration_text to event.events.

This hint cannot be replaced by an email as the deployment workflow
depends on it being in the git history. So in case a similar error
happens in the future, an additional empty commit with the omitted
deployment hint should be inserted as soon as possible. An email is a
nice touch, but only important, if there were somewhat significant
changes in the meantime.