ml: Introduce new sample user Norbert. Improve and fix tests, fix issues encountered
This introduces many backend tests for the more or less usual
susbscription workflow and fixes other tests.

This commit  introdues as new example user, Norbert, who has
only ml_admin and ml roles, which creates many unintended
situations. It is not appropiate to say this is actually
supported, see also #639.

Furthermore, this introduces an assembly list which allows
personas who do not attend to join as subscribers.