event-frontend: Improve presentation of validation of partial imports.
This at least displays the errors. It would be nicer if the location
(represented by the argname moniker which now just gives the immediate
attribute that caused the error) of the errors could be given more
precisely, but this is currently not supported by the validator. Since
this functionality should only be used by knowledable users, this
should not be huge problem.

Closes #733.