query: Replace 'similar' operators by 'match' operators.
The postgres string matching facilities have various pros and
cons. The main limitation for us is that we want to be able to
automatically search for similar variants with added diacritics. This
needs the power of regular expressions. However we can not let regexes
into the hands of unsuspecting users which will surely trip.

Previously the 'similar' query operators had regex semantics without
the necessary validation thus causing some errors. Now we disallow
a bit more, but should have weeded out the errors too.

This is a rather intrusive change, but the previous state was rather
untennable.

Closes #743