ml: provide a single entry point into the backend for all subscription actions.
For this the parameters of the action methods have been unified. They all get separate paremeters instead of a dict.

Possible actions are defined in the SubscriptionActions enum. A member of the enum is passed to the backend, where it is decided, which specific method to call for that action.

This should allow to more easily move the single methods into some external module, or even combine them all into one.