Fix some conditionals and improve tests
This commit changes the add_subscriber function to allow
adding subscribers for opt-out lists, who have unsubscribed
before. This has been a bug before, as the subscribe
function and the state transitioning graph show.
A test is added to prevent regressions here.

Furthermore, this commit does some cosmetical changes
without functional changes:

The get_interaction_policy function now returns opt_out as
interaction policy, as this represents the actual behavior
of event and assembly mailinglists better.

The write_subscription_states function does not remove
explicit subscribers who are subscribed to opt-out and
mandatory lists anymore, except if they are not allowed
to may_subscribe.
This case can never happen, but it is the same conditional
as in the add_subscriber function.