Implement select_persona for new data model
This should show any users which are not already in the list
when adding moderators, mod_subscribed and mod_unsubscribed,
and prevent you from adding people who have no means to access
the list (or have to be on it).

This does not work for moderators, who are neither assembly nor
ml admins in the assembly case. In the event case, this does not
work for moderators who are neither event admins nor ml admin or
orgas. It would be possible to change this by reading in the
rs.user.moderator set, but this might acutally be intended.