Allow a user to retrieve any info about himself
This makes a relevant difference only for past events,
as it is the only information not directly related
to the core.personas table. (The queries related to
this table are actually filtered by roles anyway.

In comparison to the behavior before 99a14cbe6744e0d,
this actually disallows orgas to see this information
if they wouldn't be able without their orga status.