ml: Add logging for automatic unsubscription
It gets kind of hard to find out for moderators why people
are suddenly gone and when they were unsubscribed from the list.
This should ease possible confusion and loss of information here,
while not spamming logs too much by using a distinct log code.

On the other hand, adding people to lists is an action that does
not undo something someone manually did earlier, so I do consider
logging of this less important.