ml: Implement mailman client functionality.
This implements the interface to the designated new mailinglist
software mailman. It utilizes the official python bindings to push the
in-database mailinglist configuration to the mailman instance running
on the mail VM.

Currently this does not try especially hard to avoid interference on
the mailman side, however the mailman self-service facilities will be
more or less counter-productive and we should probably abstain from
making the web interface accessible.

The following settings are currently applied:

* Every list is archived privately.
* HTML-mails are converted to plain text.
* To each mail a footer referencing the CdEDB is attached.
* DMARC endangered mail is wrapped inside a new mail.

This currently has the following limitations:

* Mailman does not have a way to deactivate lists, this is currently
  emulated by hiding the list and rejecting all messages sent to the
  list.
* Attachments are currently unhandled. Mailman provides some
  functionality to strip them, but these are not even exposed in the
  web interface meaning that we will have to write a custom plugin or
  something similar to get this.