frontend: Introduce frontend shards.
The idea is to make splitting up frontends easier and more
standardized.

In comparison to the variant initially chosen for the mailman
functionality this improves the readability of the code quite a
lot. The downside is, that it makes the code somewhat harder to find,
as the correct shard in which to find it is always not entirely
obvious.

fix 1

fix 1

fix 1

fix 1