frontend/event: Consider course min/max size to exclude instructors
Fixes #514

This issues changes the semantics of the event.courses.min_size and
.max_size database fields. They are now considered represent the course
instructors' limits for participants *without* counting the instructors
themselves. This is carried out on some labels in the user interface and
the logic of course size warnings.

Additionally, external tools, that deal with course sizes (e.g. course
assignment programs), should be updated to consider the fields' new
semantics in the data export.

Data of existing events that is still relevant for upcomming or past
orga teams, should be updated to reflect the new semantics.