frontend: Simplify frontend architecture.
This mainly does two things:

- Provide all backends as part of the frontend base class. They are cheap
  and it's onerous to instantiate them in each frontend class.
- Remove the per frontend session initialization. Instead all initialization
  is done centrally. This mainly affects the orga and moderator attributes
  of the request state which are now always accessible.

Additionally this renames the scriptkey to mlscriptkey to reduce namespace
pollution.

This does not make the orga and moderator attributes lazy as suggested since
that turned out to be rather tricky as it needs to work both in frontend as
well as in backend context.

Closes #870

Deploy: Fix email server configuration.
        Before deployment add new header
        echo -n "MLSCRIPTKEY: secret" > /home/listadmin/.mlscriptkey
        Afterwards delete old header