validation: provide new validator `_email_local_part` which checks for the local-part of an email address.
It would in theory be nice to deduplicate this so that the regex for this is actually the same as for
a full email, and not just a copy, but unless we really need to allow more emails this will work.

This also adjusts the `_mailinglist` validator to use `local_part` and introduces a new convention,
how to provide a list of attributes that may be present in a validated dict but will be stripped from it.

Setting `_examine_dictionary_fields` to `allow_superfluous=True` will do the same but for all keys.

This is intended for readonly attributes returned by getter methods, that cannot be set, like `address`
or `event_begin`, etc.