event: Tolerate deviations in event log upon import from offline VM.
The scenario that can happen (and has happened) is that somebody mixes up
the order of downloading an export and locking the event. In that case the
additional log message would previously cause an exception by postgres since
we do not allow to retroactively modify the log (it is append only).

Closes #971