ml/test: Fix test_no_privileges test
This test has started to fail since the refactoring
of RequestState construction introducing droid. Before
it did not work properly either, but randomly lead to
the same error expected there anyway.

However, this error is acutally only thrown if you are
trying to modify the subscription change for someone
else, which has been forgotten before.

Closes #1142