database: retroactively add changes made for the #1147 attachment versioning branch as a sql evolution file.
This MUST NOT be used in a production environment where there are actual attachments present, as it will cause existing data to be lost.

For production migration use the `bin/migrate_attachments_to_versioning.py` script instead.

This is however suitable to migrate a (local) instance where no (important) attachments are present, for testing purposes.