collation: Fix collation for Query class
This fixes the string sorting for our query pages.
Since this is the only place where we even use sql to
order stuff, sorting is finally correctly unified in
all places.

This bloats the code a bit since the conditional
treatment of COLLATE statements is to complex for
f-strings.

Closes #673

Deploy: Apply 2020-08-05_collaton.sql evolution