ml: Add default maxsize for each mailinglist type
Consulting a mailinglist table database dump, this
proposes a default maxsize for each mailinglist type.

This is given by 2 MB by default, but is heavily
restricted for mailinglist containing all members,
lowering the limit to 64 KB.
For orga lists and team lists the limit is raised to
8 MB and 4 MB respecitvely, as these users tend to
use mls in ways which do not work well with
restrictive size limits.

Closes #1167