select_persona: Allow moderators general ml user search
This has always been possible if a moderator provided the id
of any mailinglist they moderate to selectize, as it is required
for Forced/Blocked Subscriber management.

This drops the distinction between the different ml kinds and
thus allows for simpler management for multi-mailinglist pages
accessible to moderators, as for the log for all moderated
mailinglists.