cde: Rework consent decision
This commit reworks the consent decision in two ways:

First, it is now shown in the menu to all members, and extended
with a hint how to revoke a consent if given. Furthermore,
the form is only shown if the user can actually change anything
on his own.

Second, it handles the edge case that a user has obtained
searchability without deciding search more gracefully. This is
possible by admin intervention. In this case, the user can still
decide freely once, and such revoke a privacy consent given by
an admin. However, once he has decided to be searachble, he can
not revoke this anymore.

However, this still allows for the edge case that a user is
searchable with decided_search being False.

Closes #1252
Closes #1474