cde: Implement basic past course search
This is implements a course search by title, number,
description and academy which is accessible to all members.

At this point, is is still missing instructor as well as
fulltext search, but the basic functionality works.
Since this does not cover sensitive data, there are
basically no limits in contrast to member search.

Furthermore, internationalization and integration
testing is not done yet.