core: Fix purging
This commit fixes purging such that purged is an actual
user state which is distinct from being archived in the
frontend, preventing errors to happen.

Additionally, it sets all realm bits and the birthday
to -Infinity to destroy further information and to unbreak
comparisons, respectively.

Deploy: Apply 2020-10-24_purge.sql evolution

Closes #1319
Closes #1461
Closes #1338