event: Add test event with only one track and part
Such an event has been missing until now, as has been noted in #738.
Now that it is created for once, it is much easier to test that our
typical events work as intended.

I decided to set the event in the past, to simplify concluding this
event and doing purging tests, but I am not fixed on this. It could
also be in its late registration stage (or both, since we do not
validate these against each other.)