ml: Render page properly for unprivileged moderators
This hides the forms for unprivileged moderators. Furthermore, it no longer
uses mailinglist_guard to show notifications, which should never be done,
since it can be callled multiple times during a request.

Instead, this introduces a general way to add notifications in the template,
and renders them using that.

Furthermore, it heavily improves some translation.

Closes most of #1537