database: Aggresively verify privilege integrity
Since the integrity of persona privileges is a huge prerequisite
to prevent unexpected behavior, this commit adds basic checks to
make sure a basic integrity is given.

Hence, this adds an additional layer of protection to our
privilege machinery.

It could check even more, for example for purged personas, but
I think this is enough. In particular, this makes sure:

* all realms are only set if its implicit realms are set
* all members are cde users
* all admin privileges are only present if the prerequisite
  realm/admin privileges are given
* purged users are archived
* archived users are not active