notifications: Rework static notifications
Up until now, we implemented notifications indicating the page status
in one of two problematic ways: Either, we rendered them using rs.notify,
which suppressed the standard validation error message, or we put them
directly into the template. This led to them being rendered in the content
area instead of the notification area, and usually some html elements were
missing.

This commit widely deploys a solution which adds a template block in the
notification area, allowing and suggesting to render static notifications
in templates instead of the frontend. These do not interfere with validation.
However, one has to make sure not to override notifications due to template
inheritance.

Furthermore, it introduces the util.notifications macro, creating the structure
for all notification messages. It is to be used directly for static_notifications,
by custom notifications that actually need to be present in midst of a page,
and is also used by the base template to render the validation notifications.

Closes #1543