assembly/show_ballot: refactor show_own_vote makro
this was capable to work for tallied and not tallied ballots.
However, it was not used for the latter one but only displayed
after the ballot was tallied.

To the sake of simplicity (and to move it to the ballot_result page
which is only shown if the ballot is tallied), I dropped the support
of not tallied ballots.