ml: Introduce organizational ml types
This introduces two new mailinglist types for all cdedb users
and all moderators.

The former is required if there is some legal or privacy issue
relevant for all users of the cdedb, independent of their
membeership. The latter is needed if important information
about email trouble is to be sent to all moderators.

Closes #1577
Closes #970